Bugfix Model and prefix command

+++ maki/database/models/reminders.py
@@ -7,7 +7,7 @@ class Reminder(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     due_time = db.Column(db.TIMESTAMP)
     reminder = db.Column(db.String)
-    user_id = db.Column(db.BIGINT, db.ForeignKey("users.id"))
+    user_id = db.Column(db.Integer, db.ForeignKey("users.id"))
     channel_id = db.Column(db.BIGINT)
     # guild_id = db.Column(db.BIGINT, db.ForeignKey("guilds.id"))
     send_dm = db.Column(db.Boolean)
